+++ build.sbt
@@ -4,6 +4,8 @@ version := "0.1"
 
 scalaVersion := "2.12.6"
 
+scalacOptions ++= Seq("-deprecation","-feature")
+
 logBuffered in Test := false
 logLevel := Level.Warn
