+++ src/main/scala/generator/OneSide.scala
@@ -1,5 +1,8 @@
 package generator
 
-object OneSide {
+object OneSide extends Generator {
 
+  import scala.language.implicitConversions
+
+  protected implicit def transformation(polyomino: OneSide.Polyomino): List[OneSide.Polyomino] = rotations(polyomino)
 }
