+++ src/test/scala/generator/FreeSpec.scala
@@ -1,7 +1,8 @@
 package generator
 
-import generator.Utils.{Polyomino, monominos, monomino}
+//import generator.Utils.{Polyomino, monominos, monomino}
 import org.scalatest.{FlatSpec, Matchers}
+import Free._
 
 class FreeSpec extends FlatSpec with Matchers {
